requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# OCR
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# AI/ML
openai==1.3.7
langchain==0.0.350

# HTTP requests
httpx==0.25.2
aiohttp==3.9.1

# Date/Time
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
